Move pan and zoom logic to canvas
