+++ Dynavity/Dynavity/DynavityApp.swift
@@ -11,7 +11,8 @@ import SwiftUI
 struct DynavityApp: App {
     var body: some Scene {
         WindowGroup {
-            SurfaceView()
+//            SurfaceView()
+            CanvasView()
         }
     }
 }
